third_party/blink/web_tests/external/wpt/gamepad/gamepad-secure-context.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Gamepad Test: non-secure contexts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  // Old Firefox didn't expose them to non-secure contexts while Chrome and Safari always did
  test(() => {
    assert_implements(Navigator.prototype.getGamepads, "in prototype");
    assert_implements(navigator.getGamepads, "on instance");

    // Gamepad API types should also be exposed in non-secure contexts.
    assert_implements(window.Gamepad, "Gamepad interface is exposed");
    assert_implements(window.GamepadButton, "GamepadButton interface is exposed");
    assert_implements(window.GamepadEvent, "GamepadEvent interface is exposed");
  }, "Gamepad API is not exposed in non-secure contexts");
</script>
